requirements.txt
discord.py
python-dotenv
jishaku
requests
spotipy